tweak: Added hook for the recurrent pattern of loading one detour
