+++ assets/src/hooks/useLoadDetour.ts
@@ -1,0 +1,31 @@
+import { useCallback } from "react"
+import { fetchDetour } from "../api"
+import { createDetourMachine } from "../models/createDetourMachine"
+import { DetourId } from "../models/detoursList"
+import { isValidSnapshot } from "../util/isValidSnapshot"
+import { isErr } from "../util/result"
+import { useApiCall } from "./useApiCall"
+
+export const useLoadDetour = (detourId: DetourId | undefined) => {
+  const { result: detour } = useApiCall({
+    apiCall: useCallback(async () => {
+      if (detourId === undefined) {
+        return undefined
+      }
+      const detourResponse = await fetchDetour(detourId)
+      if (isErr(detourResponse)) {
+        return undefined
+      }
+      const snapshot = isValidSnapshot(
+        createDetourMachine,
+        detourResponse.ok.state
+      )
+      if (isErr(snapshot)) {
+        return undefined
+      }
+      return detourResponse.ok
+    }, [detourId]),
+  })
+
+  return detour
+}
